requirements/development.txt
-r production.txt
ipython
django-debug-toolbar
pep8
django-extensions
bs4
